Implemented display format setting
Now the user can select whether to show the full cite on top of the quote or not.

If it's enabled, It will only show the cite, without book name or chapter number only if the cite is from a single chapter. If the cite is for more than one chapter or more than one book, the book names or chapter number will be displayed.